refactor: 여러 훅과 함수로 쪼개기
