+++ src/validation/validateCvc.ts
@@ -1,0 +1,6 @@
+export const getCvcErrorMessage = (value: string) => {
+  if (value !== '' && value.length !== 3) {
+    return 'CVC는 3자리여야 합니다.';
+  }
+  return '';
+};
